Cell: refactor draggable logic - split `useDraggable()` & move it to its own file - split dragger to `CellDragger` cmp  & move it to its own file
