+++ src/styles/components.css
@@ -69,6 +69,7 @@
 @import '../components/Gradient/Gradient.css';
 @import '../components/SimpleCell/SimpleCell.css';
 @import '../components/Cell/Cell.css';
+@import '../components/CellDragger/CellDragger.css';
 @import '../components/CellButton/CellButton.css';
 @import '../components/RichCell/RichCell.css';
 @import '../components/HorizontalCell/HorizontalCell.css';
